pyproject.toml
[project]
name = "medsearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "chromadb>=1.0.12",
    "ipykernel>=6.29.5",
    "openai>=1.86.0",
    "python-dotenv>=1.1.0",
]
